Receives POST from client at ROOT/email. Parses request body, no email sending implementation yet.
